Upgrade avatax for Workarea v3
Upgrade the avatax to use the newer version 2 rest api.  They haven't
publicly released them gem so it's currently vendored into lib/avatax.
Update the tax calculator to build from price adjustments on the order
and store tax adjustments on the shipping to mirror how it works in
version 3.

AVATAX-20